Add ability to evaluate partials from templates.
Why? Well, we're starting to have documentation that is not entirely
generated from the yaml models. There's a lot of information about
custom fields that we don't want to jam into the generated code. Patials
gives us the ability to evaluate different sections of the data in
different contexts, e.g. include documentation-only markdown (actually,
full ejs templates) that exist only for the api_reference pages.