Sets findByProject and findByWorkspace to return collection
The CustomFieldSetting#findByProject and CustomField#findByWorkspace
definitions were not set to receive a collection, but the value for
`data` in these responses is an array. The implementations of these
methods (at least in the ruby client) raise errors. Setting this to
a collection should fix these issues when a new version of the libraries
are released.